backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
supabase==2.3.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
scikit-learn==1.3.2
xgboost==2.0.3
numpy==1.26.2
pandas==2.1.3
joblib==1.3.2
python-dotenv==1.0.0
